+++ disable-feature.pl
@@ -188,7 +188,7 @@ foreach $d (sort { ($b->{'alias'} ? 2 : $b->{'parent'} ? 1 : 0) <=>
 		}
 
 	# Save new domain details
-	&save_domain($d);
+	&save_domain_diff($d, $oldd);
 
 	# Run the after command
 	&set_domain_envs($d, "MODIFY_DOMAIN");